requirements.txt
discord-py-interactions
datetime
ics
requests